+++ lib/tgs/parsers/svg/builder.py
@@ -470,6 +470,8 @@ class SvgBuilder(SvgHandler, restructure.AbstractBuilder):
             svgshape = self.build_repeater(shape.lottie, shape.child, shapegroup, out_parent)
         elif isinstance(shape.lottie, objects.RoundedCorners):
             svgshape = self.build_rouded_corners(shape.lottie, shape.child, shapegroup, out_parent)
+        elif isinstance(shape.lottie, objects.Trim):
+            svgshape = self.build_trim_path(shape.lottie, shape.child, shapegroup, out_parent)
         else:
             return self.shapegroup_process_child(shape.child, shapegroup, out_parent)
         if svgshape:
@@ -515,30 +517,52 @@ class SvgBuilder(SvgHandler, restructure.AbstractBuilder):
 
     def build_rouded_corners(self, shape, child, shapegroup, out_parent):
         round_amount = shape.radius.get_value(self.time)
-        if isinstance(child, objects.Shape):
-            path = child.to_bezier()
-            bezier = path.shape.get_value(self.time).rounded(round_amount)
-            path.shape.clear_animation(bezier)
-            return self._on_shape(path, shapegroup, out_parent)
+        return self._modifier_process_child(child, shapegroup, out_parent, self._build_rouded_corners_shape, round_amount)
 
-        if isinstance(child, restructure.RestructuredShapeGroup):
-            self._build_rouded_corners_group(child, round_amount)
-            return self._on_shapegroup(child, out_parent)
+    def _build_rouded_corners_shape(self, shape, round_amount):
+        if not isinstance(shape, objects.Shape):
+            return shape
+        path = sh.to_bezier()
+        bezier = path.shape.get_value(self.time).rounded(round_amount)
+        path.shape.clear_animation(bezier)
+        return path
 
-        return child
+    def build_trim_path(self, shape, child, shapegroup, out_parent):
+        start = shape.start.get_value(self.time) / 100
+        end = shape.end.get_value(self.time) / 100
 
-    def _build_rouded_corners_group(self, shapegroup, round_amount):
+        return self._modifier_process_child(child, shapegroup, out_parent, self._build_trim_path_shape, start, end)
+
+    def _build_trim_path_shape(self, shape, start, end):
+        if not isinstance(shape, objects.Shape):
+            return shape
+        path = shape.to_bezier()
+        bezier = path.shape.get_value(self.time).segment(start, end)
+        path.shape.clear_animation(bezier)
+        return path
+
+    def _modifier_process_children(self, shapegroup, out_parent, callback, *args):
         children = []
-        for sh in shapegroup.children:
-            if isinstance(sh, objects.Shape):
-                path = sh.to_bezier()
-                bezier = path.shape.get_value(self.time).rounded(round_amount)
-                path.shape.clear_animation(bezier)
-                sh = path
-            elif isinstance(sh, restructure.RestructuredShapeGroup):
-                self._build_rouded_corners_group(sh, round_amount)
-            children.append(sh)
+        for shape in shapegroup.children:
+            children.append(self._modifier_process_child(shape, shapegroup, out_parent, callback, *args))
         shapegroup.children = children
+
+    def _modifier_process_child(self, shape, shapegroup, out_parent, callback, *args):
+        if isinstance(shape, restructure.RestructuredShapeGroup):
+            self._modifier_process_children(shape, out_parent, callback, *args)
+            return self._on_shapegroup(shape, out_parent)
+        elif isinstance(shape, restructure.RestructuredPathMerger):
+            shape.paths = [
+                callback(p, *args)
+                for p in shape.paths
+            ]
+            return self._on_merged_path(shape, shapegroup, out_parent)
+        #elif isinstance(shape, restructure.RestructuredModifier):
+            #return self._on_shape_modifier(shape, shapegroup, out_parent)
+        else:
+            shape = callback(shape, *args)
+            return self._on_shape(shape, shapegroup, out_parent)
+
 
     def _custom_object_supported(self, shape):
         if has_font and isinstance(shape, font.FontShape):