+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
  * Implemented fill rule
+ * Partial support for text layers on SVG export
 
 
 ## 0.5.1 - 2020-04-19