+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Implemented fill rule
  * Partial support for text layers on SVG export
  * Support solid color layers on SVG export
+ * Partial support for trim paths on SVG export
 
 
 ## 0.5.1 - 2020-04-19