+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Partial support for text layers on SVG export
  * Support solid color layers on SVG export
  * Partial support for trim paths on SVG export
+ * Added support for dotLottie
 
 
 ## 0.5.1 - 2020-04-19